fix(ci): round-3 Rust Nightly — fmt packing, clippy lint sweep, rustdoc links
Apply rustfmt 1.97 exact diffs (4 files: main.rs import packing, channels.rs
split, gateway sandbox.rs collapse, scheduler delivery.rs collapse), restore
Tool/ToolResult imports for tools lib tests, gate webhook BlockEncrypt to the
test module, and fix ~66 clippy lints across skills/channels/session plus
rustdoc link errors in memory/skills/channels and session doc-list indentation.

+++ crates/session/src/manager.rs
@@ -1334,7 +1334,7 @@ impl SessionManager {
     /// `Active`. Returns the number of sessions resumed.
     pub fn recover_interrupted_compactions(&self) -> CoreResult<usize> {
         let report = self.recover(false, chrono::Duration::days(30))?;
-        Ok(report.interrupted_compactions as usize)
+        Ok(report.interrupted_compactions)
     }
 
     // -----------------------------------------------------------------------
